Improved error handling
The EvError enum (new name, so as not to clash with std::err::Error)
contains more items that better explains the nature of some
errors (e.g., too many sides or too many dice).

The number of dice and sides and the bonus/malus now use 16-bit
integers; I figure that anything larger is over-doing it.